+++ src/app/manage/[objects]/[action]/page.tsx
@@ -1,6 +1,6 @@
 import { Header } from '@/src/components/common/header';
 import { Footer } from '@/src/components/common/footer';
-import { ObjectContainer } from '@/src/app/management/[objects]/[action]/object-container';
+import { ObjectContainer } from '@/src/app/manage/[objects]/[action]/object-container';
 
 export default function ObjectManagementPage() {
     return (